добавил unit тесты

+++ app/src/main/java/com/aka_npou/sberandroidschool_finalproject/data/entity/TotalStatisticEntity.java
@@ -18,7 +18,7 @@ public class TotalStatisticEntity {
     }
 
     @Ignore
-    public TotalStatisticEntity(int countQuestions, int countCorrectAnswers, int firstDay) {
+    public TotalStatisticEntity(int countQuestions, int countCorrectAnswers, long firstDay) {
         this.countQuestions = countQuestions;
         this.countCorrectAnswers = countCorrectAnswers;
         this.firstDay = firstDay;
